fix ui on mobile, delete permission and coding standards.

+++ gp-includes/routes/note.php
@@ -24,6 +24,11 @@ class GP_Route_Note extends GP_Route_Main {
 	public function new_post() {
 		$translation_id = gp_post( 'translation_id' );
 		$note           = gp_post( 'note' );
+		$translation    = new GP_Translation( array( 'id' => $translation_id ) );
+
+		if ( ! GP::$permission->current_user_can( 'approve', 'translation', $translation_id, array( 'translation' => $translation ) ) ) {
+			return false;
+		}
 
 		if ( ! $this->verify_nonce( 'new-note-' . $translation_id ) ) {
 			return $this->die_with_error( __( 'An error has occurred. Please try again.', 'glotpress' ), 403 );
@@ -31,9 +36,11 @@ class GP_Route_Note extends GP_Route_Main {
 
 		$this->notices[] = __( 'The note was created!', 'glotpress' );
 		$translation     = GP::$translation->get( $translation_id );
-		$note_object     = GP::$notes->save();
+		$note            = GP::$notes->save();
 
-		return $this->render_note( $note_object, $translation );
+		$this->tmpl( 'note', get_defined_vars() );
+
+		return $note;
 	}
 
 	/**
@@ -47,6 +54,11 @@ class GP_Route_Note extends GP_Route_Main {
 		$translation_id = gp_post( 'translation_id' );
 		$note           = gp_post( 'note' );
 		$note_id        = gp_post( 'note_id' );
+		$translation    = new GP_Translation( array( 'id' => $translation_id ) );
+
+		if ( ! GP::$permission->current_user_can( 'approve', 'translation', $translation_id, array( 'translation' => $translation ) ) ) {
+			return false;
+		}
 
 		if ( ! $this->verify_nonce( 'edit-note-' . $note_id ) ) {
 			return $this->die_with_error( __( 'An error has occurred. Please try again.', 'glotpress' ), 403 );
@@ -54,9 +66,9 @@ class GP_Route_Note extends GP_Route_Main {
 
 		$this->notices[] = __( 'The note was updated!', 'glotpress' );
 		$translation     = GP::$translation->get( $translation_id );
-		$note_object     = GP::$notes->edit( $note_id, $note, $translation );
+		$note            = GP::$notes->edit( $note_id, $note, $translation );
 
-		return $this->render_note( $note_object, $translation );
+		$this->tmpl( 'note', get_defined_vars() );
 	}
 
 	/**
@@ -67,27 +79,21 @@ class GP_Route_Note extends GP_Route_Main {
 	 * @return bool
 	 */
 	public function delete_post() {
-		$note_id = gp_post( 'note_id' );
+		$translation_id = gp_post( 'translation_id' );
+		$note_id        = gp_post( 'note_id' );
+		$translation    = new GP_Translation( array( 'id' => $translation_id ) );
 
 		if ( ! $this->verify_nonce( 'delete-note-' . $note_id ) ) {
 			return $this->die_with_error( __( 'An error has occurred. Please try again.', 'glotpress' ), 403 );
+		}
+
+		if ( ! GP::$permission->current_user_can( 'approve', 'translation', $translation_id, array( 'translation' => $translation ) ) ) {
+			return false;
 		}
 
 		$this->notices[] = __( 'The note was deleted!', 'glotpress' );
 		GP::$notes->delete_all( array( 'id' => $note_id ) );
 
 		return true;
-	}
-
-	/**
-	 * Render the note
-	 *
-	 * @param object $note        Note object.
-	 * @param object $translation Translation object.
-	 */
-	private function render_note( $note, $translation ) {
-		require_once GP_TMPL_PATH . 'helper-functions.php';
-		$can_approve = $this->can( 'approve', 'translation-set', $translation->translation_set_id );
-		return render_note( $note, $can_approve );
 	}
 }
